+++ .php-cs-fixer.dist.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+$finder = (new PhpCsFixer\Finder())
+    ->in([__DIR__.'/src', __DIR__.'/tests', __DIR__.'/tools'])
+    ->exclude('php-cs-fixer');
+
+return (new PhpCsFixer\Config())
+    ->setRiskyAllowed(true)
+    ->setRules([
+        '@PSR12' => true,
+        '@PSR12:risky' => true,
+        'concat_space' => ['spacing' => 'none'],
+        'declare_strict_types' => true,
+        'native_function_invocation' => false,
+        'no_unused_imports' => true,
+        'ordered_imports' => ['sort_algorithm' => 'alpha'],
+        'single_quote' => true,
+        'trailing_comma_in_multiline' => ['elements' => ['arguments', 'arrays', 'parameters']],
+        'global_namespace_import' => ['import_classes' => false, 'import_constants' => false, 'import_functions' => false],
+    ])
+    ->setFinder($finder);